Renamed service to a more generic name
