+++ api-gateway/src/main/resources/application.yml
@@ -8,10 +8,10 @@ spring:
           enabled: true
           lower-case-service-id: true # not to use uppercase: localhost:8080/MY-SERVICE/... but lowercase instead
       routes:
-        - id: lang-service
-          uri: lb://lang-service
+        - id: example-service
+          uri: lb://example-service
           predicates:
-            - Path=/api/person/**
+            - Path=/api/example/**
           filters:
             - RewritePath=/api(?<path>/?.*), $\{path}
         - id: sample-service
