+++ example-service/src/main/resources/application-dev.yml
