+++ example-service/src/main/resources/bootstrap.yml
