+++ sample-service/src/main/kotlin/com/horakm/sample/client/ExampleServiceClient.kt
@@ -3,8 +3,8 @@ package com.horakm.sample.client
 import org.springframework.cloud.openfeign.FeignClient
 import org.springframework.web.bind.annotation.GetMapping
 
-@FeignClient(name = "lang-service", path="/person")
-interface LangServiceClient {
+@FeignClient(name = "example-service", path="/example/person")
+interface ExampleServiceClient {
     @GetMapping("/property1")
     fun getProperty(): String
 }
