+++ sample-service/src/main/kotlin/com/horakm/sample/controller/SampleController.kt
@@ -1,6 +1,6 @@
 package com.horakm.sample.controller
 
-import com.horakm.sample.client.LangServiceClient
+import com.horakm.sample.client.ExampleServiceClient
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/sample")
-class SampleController(private val langServiceClient: LangServiceClient) {
+class SampleController(private val exampleServiceClient: ExampleServiceClient) {
 
     @Value("\${property1}")
     lateinit var prop1: String
@@ -17,6 +17,6 @@ class SampleController(private val langServiceClient: LangServiceClient) {
     fun getProperty(): String = prop1
 
     @GetMapping("/externalProperty")
-    fun getClientProperty(): String = langServiceClient.getProperty()
+    fun getClientProperty(): String = exampleServiceClient.getProperty()
 
 }
